[GOHW-5]: Добавлены тесты для рейт лимитера

+++ cart/internal/app/config.go
@@ -19,7 +19,7 @@ type Config struct {
 			RetryStatusCodes []int `json:"RetryStatusCodes"`
 			MaxRetries       int   `json:"MaxRetries"`
 		} `json:"RetryPolicy"`
-		RPS int `json:"RPS"`
+		RPS uint64 `json:"RPS"`
 	} `json:"ProductService"`
 	LOMS struct {
 		Address string `json:"Address"`
